backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-speech==2.21.0
google-cloud-language==2.11.0
firebase-admin==6.2.0
tensorflow==2.15.0
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
pydub==0.25.1
soundfile==0.12.1